add query parameter management functions
