+++ webapp/pages_/home.py
@@ -4,11 +4,12 @@ import os
 
 import streamlit as st
 from service.components import show_sandbox_message
+from service.query_params import get_all_query_params
 from service.utils import APP_URL, _log, display_info_message, quanting_output_path
 
 from shared.keys import FORBIDDEN_CHARACTERS_IN_RAW_FILE_NAME, EnvVars
 
-_log(f"loading {__file__} {st.query_params}")
+_log(f"loading {__file__} {get_all_query_params()}")
 
 st.set_page_config(page_title="AlphaKraken: home", layout="wide")
